+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.1.0
+
+Simplified the API for creating related models, assuming you are using Factory Discovery or have mapped your factories
+already. If you have, you can now call `new PostFactory().With<Post>(p => p.Post)`. In other words, you no longer need
+to supply a generic argument for the factory type.
+
 # 1.0.1
 
 Target framework is now dotnet 6.0